+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to `filament-page-builder` will be documented in this file.
 
+## 3.0.0 Initial filament v3 update
+- support Filamentphp v3 / Livewire v3
+- disable validation for preview
+- visual fixes for preview
+
+Please note there is a bug, preventing you from saving blocks, in filament `Builder` in combination with the `filament/spatie-laravel-translatable-plugin`
+https://github.com/filamentphp/filament/issues/8656
+
 ## 1.0.0 - 202X-XX-XX
 
 - initial release